MPSKit.jl documentation

This code track contains the numerical research and development of the Ghent Quantum Group with regard to tensor network simulation in the julia language. The purpose of this package is to facilitate efficient collaboration between different members of the group.

Topics of research on tensor networks within the realm of this track include:

Tensor network algorithms (excitations, tdvp, vumps, ...)

MPS routines (MPS diagonalization, Schmidt Decomposition, MPS left and right multiplication, ...)

The study of several useful models (nearest neighbour interactions, MPO's, long range interactions, ...)

Contents

MPSKit.jl documentation

Contents

Statement of Intent

Getting Started

Functions

Supporting and Credit

Statement of Intent

Quantum mechanics is one of the biggest breakthroughs in the history of physics as, for instance, the successful applications of the theory to the hydrogen atom showed. Currently, almost a century later, physicists are still struggling to fully understand the implications of quantum mechanics. Especially in systems with many degrees of freedom where the complexity escalates quickly. In recent years, it was realized that quantum entanglement can be exploited to create an efficient ansatz for the ground and low energy states of a many-body system. These ansatz are known as ‘tensor network states’.

Tensor network states allow for a more efficient use of time and computational resources since one no longer has to look for the correct answer in the entire space of possible states, restricting to the physical subspace of tensor network states is sufficient. The local structure of these states also provide ways to get a deeper understanding of generic emergent phenomena in quantum many-body systems and their connection to entanglement.

Hence, tensor network techniques are finding their way in an increasing number of fields in physics, ranging from condensed matter theory to quantum gravity. In the UGent quantum group of Prof. Verstraete and Prof. Haegeman we are at the forefront of many of these applications, with an expertise in both the theoretical and numerical aspects. With this code base we aim to create a flexible, well-optimized and inclusive tool to support and stimulate rapid expansion of this expertise.

Getting Started

For those who want to dive in immediately, the examples folder is a good starting point. Those who want read up first can use this documentation file.

Functions

MPSKit.Dmrg — Type

onesite dmrg

MPSKit.Dmrg2 — Type

twosite dmrg

MPSKit.FiniteMpo — Type

FiniteMps(data::Array)

finite one dimensional mps
algorithms usually assume a right-orthormalized input

MPSKit.FiniteMps — Type

FiniteMps(data::Array)

finite one dimensional mps
algorithms usually assume a right-orthormalized input

MPSKit.Idmrg1 — Type

onesite infinite dmrg

MPSKit.Idmrg2 — Type

twosite infinite dmrg

MPSKit.MpoHamiltonian — Type

MpoHamiltonian

represents a general periodic quantum hamiltonian

MPSKit.MpsCenterGauged — Type

st = MpsCenterGauged(a::Array)

Type definition of a uniform center gauged mps.
st.AL[i]*st.CR[i] == st.CR[i-1]*st.AR[i] == st.AC[i]
st.AL[i] is left unitary
st.AR[i] is right unitary

MPSKit.MpsComoving — Type

MpsComoving(leftstate,window,rightstate)

muteable window of tensors on top of an infinite chain

MPSKit.MpsMultiline — Type

2d extension of MpsCenterGauged

MPSKit.OptimalExpand — Type

expands the given mps using the algorithm given in the vumps paper

MPSKit.RandExpand — Type

use a random matrix to expand ...

MPSKit.SimpleManager — Type

a very simple bond dimension manager when manager.criterium(state) is satisfied it does nothing if not, it will use algorithm manager.change to change the bonds

MPSKit.SvdCut — Type

Truncate a given state using svd

MPSKit.Tdvp — Type

onesite tdvp

MPSKit.Tdvp2 — Type

twosite tdvp (works for finite mps's)

MPSKit.UnionTrunc — Type

Take the union of 2 truncation algorithms using &

MPSKit.Vumps — Type

see https://arxiv.org/abs/1701.07035

MPSKit.VumpsSvdCut — Type

use an idmrg2 step to truncate/expand the bond dimension

MPSKit.dynamicaldmrg — Method

https://arxiv.org/pdf/cond-mat/0203500.pdf

MPSKit.entropy — Method

calculates the entropy of a given state

MPSKit.exp_decomp — Method

Decomposes fun into a sum of exponentials (coeff,exponents)=exp_decomp(fun;fitdist,numexp);

MPSKit.find_groundstate — Function

find_groundstate(state,ham,alg,pars=params(state,ham))

find the groundstate for ham using algorithm alg

MPSKit.l_LL — Method

l_LL(state,location)
Left dominant eigenvector of the AL-AL transfermatrix

MPSKit.l_LR — Function

l_LR(state,location)
Left dominant eigenvector of the AL-AR transfermatrix

MPSKit.l_RL — Function

l_RL(state,location)
Left dominant eigenvector of the AR-AL transfermatrix

MPSKit.l_RR — Function

l_RR(state,location)
Left dominant eigenvector of the AR-AR transfermatrix

MPSKit.leading_boundary — Function

leading_boundary(state,ham,alg,pars=params(state,ham))

approximate the leading eigenvector for ham

MPSKit.r_LL — Function

r_LL(state,location)
Right dominant eigenvector of the AL-AL transfermatrix

MPSKit.r_LR — Function

r_LR(state,location)
Right dominant eigenvector of the AL-AR transfermatrix

MPSKit.r_RL — Function

r_RL(state,location)
Right dominant eigenvector of the AR-AL transfermatrix

MPSKit.r_RR — Method

r_RR(state,location)
Right dominant eigenvector of the AR-AR transfermatrix

MPSKit.spinmatrices — Method

Returns spin operators Sx,Sy,Sz,Id for spin s

MPSKit.splitham — Method

given a thermal state, you can map it to an mps by fusing the physical legs together to prepare a gibbs ensemble, you need to evolve this state with H working on both legs here we return the 'superhamiltonian' (Hid,idH)

MPSKit.timestep — Function

(newstate,newpars) = timestep(state,hamiltonian,dt,alg,pars = params(state,hamiltonian))

evolves state forward by dt using algorithm alg

Index

This is the index.

MPSKit.Dmrg

MPSKit.Dmrg2

MPSKit.FiniteMpo

MPSKit.FiniteMps

MPSKit.Idmrg1

MPSKit.Idmrg2

MPSKit.MpoHamiltonian

MPSKit.MpsCenterGauged

MPSKit.MpsComoving

MPSKit.MpsMultiline

MPSKit.OptimalExpand

MPSKit.RandExpand

MPSKit.SimpleManager

MPSKit.SvdCut

MPSKit.Tdvp

MPSKit.Tdvp2

MPSKit.UnionTrunc

MPSKit.Vumps

MPSKit.VumpsSvdCut

MPSKit.dynamicaldmrg

MPSKit.entropy

MPSKit.exp_decomp

MPSKit.find_groundstate

MPSKit.l_LL

MPSKit.l_LR

MPSKit.l_RL

MPSKit.l_RR

MPSKit.leading_boundary

MPSKit.r_LL

MPSKit.r_LR

MPSKit.r_RL

MPSKit.r_RR

MPSKit.spinmatrices

MPSKit.splitham

MPSKit.timestep

Supporting and Credit

If you have any questions, bug reports, feature requests, etc., please submit an issue on github. If you would like to contribute to the moral of the coding team, please send us food. The software in this stack was developed as part of academic research of the UGent Quantum Group of Prof. Verstraete and Prof. Haegeman.